Add reroll button
Sampling is non-deterministic, but seed isn't exposed. Properly fixing
this is a high priority, but until I figure it out, a reroll button has
been added as a temporary fix that will mark the node as stale.

Removed the cast from model load. It seems to force a copy which slows
down the initial load.

Remove an errant space from the README